test/send_recv: fix vectorized send testing for 32-bit builds
The test uses unsigned long, which is still 4 bytes in size on 32-bit
builds. This causes verification failures:

Running test send_recv.t
data mismatch at 512: 257
test small vec sync failed
Test send_recv.t failed with ret 1

Switch it to use uint64_t, making it the same size on 32-bit and 64-bit
builds.

Link: https://github.com/axboe/liburing/issues/1501
Fixes: cb2aaf3b7ead ("test/send_recv: add support for testing IORING_SEND_VECTORIZED")

+++ test/send_recv.c
@@ -7,6 +7,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <inttypes.h>
 #include <arpa/inet.h>
 #include <sys/types.h>
 #include <sys/socket.h>
@@ -17,7 +18,7 @@
 
 #define MAX_MSG	4096
 
-static unsigned long str[MAX_MSG / sizeof(unsigned long)];
+static uint64_t str[MAX_MSG / sizeof(uint64_t)];
 
 #define PORT	10202
 #define HOST	"127.0.0.1"
@@ -88,7 +89,7 @@ err:
 static int do_recv(struct io_uring *ring, struct iovec *iov, int enobufs)
 {
 	struct io_uring_cqe *cqe;
-	unsigned long *ptr;
+	uint64_t *ptr;
 	int i, ret;
 
 	ret = io_uring_wait_cqe(ring, &cqe);
@@ -118,10 +119,10 @@ static int do_recv(struct io_uring *ring, struct iovec *iov, int enobufs)
 	}
 
 	ptr = iov->iov_base;
-	for (i = 0; i < MAX_MSG / sizeof(unsigned long); i++) {
+	for (i = 0; i < MAX_MSG / sizeof(uint64_t); i++) {
 		if (ptr[i] == str[i])
 			continue;
-		fprintf(stderr, "data mismatch at %d: %lu\n", i, ptr[i]);
+		fprintf(stderr, "data mismatch at %d: %llu\n", i, (unsigned long long) ptr[i]);
 		goto err;
 	}
 
@@ -224,7 +225,7 @@ retry:
 	sqe = io_uring_get_sqe(&ring);
 	if (vec) {
 		size_t total = MAX_MSG;
-		unsigned long *ptr = str;
+		uint64_t *ptr = str;
 		int i, nvecs;
 
 		if (!big_vec) {
@@ -239,7 +240,7 @@ retry:
 			for (i = 0; i < 32; i++) {
 				vecs[i].iov_base = ptr;
 				vecs[i].iov_len = total;
-				ptr += total / sizeof(unsigned long);
+				ptr += total / sizeof(uint64_t);
 			}
 			nvecs = 32;
 		}
@@ -371,7 +372,7 @@ int main(int argc, char *argv[])
 	if (argc > 1)
 		return T_EXIT_SKIP;
 
-	for (i = 0; i < MAX_MSG / sizeof(unsigned long); i++)
+	for (i = 0; i < MAX_MSG / sizeof(uint64_t); i++)
 		str[i] = i + 1;
 
 	ret = test_invalid();
